dict: change id type from uint8_t to int
This is needed for bitfields longer than 8 bit (ie. ADS-C v2
VerticalType)

+++ src/util.c
@@ -47,7 +47,7 @@ void *xrealloc(void *ptr, size_t size, const char *file, const int line, const c
 	return ptr;
 }
 
-void *dict_search(const dict *list, uint8_t id) {
+void *dict_search(const dict *list, int id) {
 	if(list == NULL) return NULL;
 	dict *ptr;
 	for(ptr = (dict *)list; ; ptr++) {
